develop/ai-tools/backend/pyproject.toml
[build-system]
requires = ["setuptools>=78.1.1", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "Medical Agentic AI Research Subsystem"
version = "0.1.0"
description = "An agentic, LangGraph-driven medical research assistant that fuses RAG (101GenAI) and Google CSE to deliver concise, cited answers from trusted sources for any medical question"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Naveen Putta, David Marshall, Ravi Kumar"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Core dependencies
    "fastapi>=0.110.0",
    "uvicorn>=0.28.0",
    "httpx>=0.27.0",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "python-dotenv>=1.1.1",
    "annotated-types>=0.7.0",
    "typing-inspection>=0.4.1",
    "typing-extensions>=4.14.1",
    
    # LangGraph & AI
    "langgraph>=0.2.2",
    "google-generativeai>=0.6.0",
    "google-api-python-client>=2.117.0",
    
    # Web processing & requests
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=4.9.3",
    "aiohttp>=3.9.3",
    
    # Utilities
    "tenacity>=8.2.3",
    "cachetools>=5.3.2",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "isort>=5.13.2",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[project.urls]
"Documentation" = "https://coligomed.atlassian.net/wiki/x/CwA-Pw"
"Source Code" = "https://gitlab.coligomed.com/DHP/dhp-backend-service/ai-tools.git"
"Issue Tracker" = ""

[tool.setuptools]
packages = ["api", "models", "routes", "utils", "workflows"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--strict-markers --color=yes"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[project.scripts]
ai-tools-server = "app:main"
rag-server = "Rag_server:main"
websearch-server = "websearch_server:main"
